mpesa integration

+++ gms/services/payments.py
@@ -6,13 +6,12 @@ from gms.services.utils import format_mobile_number
 from gms.services.utils import create_payment_entry
 from gms.services.utils import send_sms
 from datetime import datetime
-from gms.services.update import update_mgr_status, enqueue_member_contribution, update_table_banking_fund, process_loan_repayment
 
 
 
 
 class ProcessPayment:
-    def safaricom_stk_push_request(self, payload):
+    def safaricom_stk_push_request(self, context):
         try:
             url = "https://api.safaricom.co.ke/mpesa/stkpush/v1/processrequest"
 
@@ -21,12 +20,12 @@ class ProcessPayment:
                 "Password": f"{ self.get_env_data()['PASSWORD'] }",
                 "Timestamp": "20240514183146",
                 "TransactionType": "CustomerPayBillOnline",
-                "Amount": f"{payload['amount']}",
-                "PartyA": f"{payload['mobile_number']}",
-                "PartyB": f"{payload['paybill']}",
-                "PhoneNumber": f"{payload['mobile_number']}",
-                "CallBackURL": "https://7895-102-212-236-161.ngrok-free.app/api/method/gms.services.payments.stk_push_response",
-                "AccountReference": f"{payload['bank_account_number']}",
+                "Amount": f"{context['amount']}",
+                "PartyA": f"{context['mobile_number']}",
+                "PartyB": f"{context['paybill']}",
+                "PhoneNumber": f"{context['mobile_number']}",
+                "CallBackURL": "https://52f3-102-213-179-89.ngrok-free.app/api/method/gms.services.payments.stk_push_response",
+                "AccountReference": f"{context['bank_account_number']}",
                 "TransactionDesc": "Gym Payment"
             }
             
@@ -85,7 +84,7 @@ class ProcessPayment:
 payment = ProcessPayment()
 
 
-@frappe.whitelist(methods=['GET'], allow_guest=True)
+@frappe.whitelist( allow_guest=True)
 def make_payment(amount, mobile_number, invoice_number):
     try:
         existing_transaction_name = frappe.db.get_all(
@@ -98,14 +97,17 @@ def make_payment(amount, mobile_number, invoice_number):
             fields=['name']
         )
         
-
-        for transaction in existing_transaction_name:
-            frappe.db.delete("Payment Transaction", {
-                "name": transaction.name,
-            })
+        if existing_transaction_name:
+            for transaction in existing_transaction_name:
+                frappe.db.delete("Payment Transaction", {
+                    "name": transaction.name,
+                })
         
         frappe.db.commit()
-        paybill, account_number = frappe.db.get_value('GMS Payment Account', None, ['paybill', 'account_number'])
+        payment_account = frappe.get_single("GMS Payment Account")
+        paybill = payment_account.paybill
+        account_number = payment_account.account_number
+    
 
         doc = frappe.get_doc({
             'doctype': 'Payment Transaction',
@@ -136,17 +138,11 @@ def make_payment(amount, mobile_number, invoice_number):
 @frappe.whitelist(allow_guest=True)
 def stk_push_response():
     # data = frappe.request.get_json(force=True)
-
-    data={'Body': {'stkCallback': {'MerchantRequestID': '4f9d-4622-a0da-1c77977dad0c137103410', 
-    'CheckoutRequestID': 'ws_CO_11112024114458752768135284', 'ResultCode': 0,
-     'ResultDesc': 'The service request is processed successfully.',
-     'CallbackMetadata': {'Item': [{'Name': 'Amount', 'Value': 1.0}, {'Name': 'MpesaReceiptNumber', 'Value': 'SKB5EAPG87'}, {'Name': 'Balance'}, {'Name': 'TransactionDate', 'Value': 20241111114510}, {'Name': 'PhoneNumber', 'Value': 254768135284}]}}}}
-
-
+    data={'Body': {'stkCallback': {'MerchantRequestID': '60e4-4f14-997e-f04c8c4f586d142963096', 'CheckoutRequestID': 'ws_CO_12112024122212507768135284', 'ResultCode': 0, 'ResultDesc': 'The service request is processed successfully.', 'CallbackMetadata': {'Item': [{'Name': 'Amount', 'Value': 1.0}, {'Name': 'MpesaReceiptNumber', 'Value': 'SKC2ILC7T0'}, {'Name': 'Balance'}, {'Name': 'TransactionDate', 'Value': 20241112122236}, {'Name': 'PhoneNumber', 'Value': 254768135284}]}}}}
+    
     response = data['Body']['stkCallback'] 
 
     items = response['CallbackMetadata']['Item']
-    
 
     Amount = next((item["Value"] for item in items if item["Name"] == "Amount"), None)
     MpesaReceiptNumber = next((item["Value"] for item in items if item["Name"] == "MpesaReceiptNumber"), None)
@@ -155,8 +151,14 @@ def stk_push_response():
 
     reference_code = response['ResultCode']
     
-    
     try:
+        docs=frappe.get_all("MPesa Payment Transaction",{"mpesa_receipt_number":MpesaReceiptNumber},{"name"})
+        if docs:
+            for item in docs:
+                frappe.db.delete("MPesa Payment Transaction", {
+                    "name": item.name,
+                })
+            frappe.db.commit()
         doc = frappe.get_doc({
             'doctype': 'MPesa Payment Transaction',
             'merchant_request_id': response['MerchantRequestID'],
@@ -180,24 +182,8 @@ def stk_push_response():
 
 
 @frappe.whitelist(allow_guest=True)
-def process_payment_and_reconcile_member_invoice():
+def process_payment_and_reconcile_member_invoice(MpesaReceiptNumber, TransactionDate, PhoneNumber, Amount):
     try:
-        data={'Body': {'stkCallback': {'MerchantRequestID': '4f9d-4622-a0da-1c77977dad0c137103410', 
-        'CheckoutRequestID': 'ws_CO_11112024114458752768135284', 'ResultCode': 0,
-        'ResultDesc': 'The service request is processed successfully.',
-        'CallbackMetadata': {'Item': [{'Name': 'Amount', 'Value': 1.0}, {'Name': 'MpesaReceiptNumber', 'Value': 'SKB5EAPG87'}, {'Name': 'Balance'}, {'Name': 'TransactionDate', 'Value': 20241111114510}, {'Name': 'PhoneNumber', 'Value': 254768135284}]}}}}
-
-
-        response = data['Body']['stkCallback'] 
-
-        items = response['CallbackMetadata']['Item']
-        
-
-        Amount = next((item["Value"] for item in items if item["Name"] == "Amount"), None)
-        MpesaReceiptNumber = next((item["Value"] for item in items if item["Name"] == "MpesaReceiptNumber"), None)
-        TransactionDate = next((item["Value"] for item in items if item["Name"] == "TransactionDate"), None)
-        PhoneNumber = next((item["Value"] for item in items if item["Name"] == "PhoneNumber"), None)
-
         if not isinstance(PhoneNumber, str):
             PhoneNumber = str(PhoneNumber)
 
@@ -210,9 +196,11 @@ def process_payment_and_reconcile_member_invoice():
         )
                 
         if not pending_transaction:
-            return
+            return "no pending transaction"
 
-        paid_to_account = frappe.db.get_value("Mode of Payment Account", {"parent": "M-Pesa"}, "default_account")
+        all=frappe.get_all("Mode of Payment Account",{"parent":"M-Pesa"},"default_account")
+        if all:
+            paid_to_account = all[0].default_account
 
         payment_reference = [{
             "reference_doctype": "Sales Invoice",
@@ -239,55 +227,19 @@ def process_payment_and_reconcile_member_invoice():
         
         invoice_doc = frappe.get_doc("Sales Invoice", pending_transaction[0].invoice_number)
 
-        if invoice_doc.custom_contribution_type:
-            member_name = frappe.db.get_value("Customer", {"name": customer}, "custom_chamaa_member")
+       
+    
 
-            if member_name:
-                current_month = datetime.now().strftime("%B")
-                month_to_use = invoice_doc.custom_invoice_month if invoice_doc.custom_invoice_month else current_month
-                
-                enqueue_member_contribution(
-                    member=member_name,
-                    amount=Amount,
-                    contribution_type=invoice_doc.custom_contribution_type,
-                    month=month_to_use,  
-                    date_of_contribution=TransactionDate,
-                    phone_number=standardized_phone_number,
-                    invoice_name=pending_transaction[0].invoice_number, 
-                    merry_go_round_number=invoice_doc.custom_merry_go_round_number,  
-                    table_banking_fund_number=invoice_doc.custom_table_banking_number 
-                )
-
-        if invoice_doc.custom_contribution_type == "Merry Go Round":
-            round_number = invoice_doc.custom_merry_go_round_number  
-            update_mgr_status(round_number, pending_transaction[0].invoice_number, Amount) 
-
-        elif invoice_doc.custom_contribution_type == "Table Banking Fund":
-            table_banking_fund_number = invoice_doc.custom_table_banking_number
-            update_table_banking_fund(
-                member=member_name,
-                amount=Amount,
-                date_of_contribution=TransactionDate,
-                table_banking_fund_number=table_banking_fund_number
-            )
-        elif invoice_doc.custom_contribution_type == "Loan Repayment":
-            loan_name = frappe.db.get_value('Loan', {'loan_invoice': invoice_doc.name}, 'name')
-            if loan_name:
-                process_loan_repayment(
-                    loan_name=loan_name,
-                    amount_to_pay=Amount,
-                    invoice_name=invoice_doc.name
-                )
+     
 
         frappe.db.commit()
 
-        full_name = frappe.db.get_value('Member', {"name": member_name}, "full_name")
-        message = f" Dear {full_name}, Thank you for your {invoice_doc.custom_contribution_type} contribution of {Amount}. It has been successfully received."
+        message = f" Dear {customer}, Thank you for your gms payment of {Amount}. It has been successfully received."
         send_sms(standardized_phone_number, message)
 
         return {
             'success': True,
-            'message': 'Payment reconciled and Member Contribution created successfully',
+            'message': 'Payment reconciled  with locker booking successfully',
             'invoice_number': pending_transaction[0].invoice_number
         }
